updating types of variables

+++ src/util.c
@@ -13,7 +13,7 @@ void velocity_gen(double *x, double *y, int type) {
 		return;
 	}
 
-	max_v = (type == LIGHT) ? velocityLightMin : 
+	max_v = (type == LIGHT) ? velocityLightMin :
 					(type == MEDIUM) ? velocityMediumMin :
 					(type == HEAVY) ? velocityHeavyMin : 0;
 	min_v = (type == LIGHT) ? velocityLightMax :
@@ -57,7 +57,7 @@ void print_properties_h(void) {
 	LOG(("  %f <=   mass   <= %f\n", massHeavyMin, massHeavyMax));
 }
 
-void print_particle(double *p, double time) {
+void print_particle(double *p, int time) {
 /*
 	LOG(("{\n"));
 	LOG(("\tparticle_id: %f,\n", p[ID_COL]));
@@ -70,7 +70,7 @@ void print_particle(double *p, double time) {
 	LOG(("}\n"));
 */
 /*
-	LOG(("t=%f, p_id=%d, mass=%f, p_x=%f, p_y=%f, v_x=%f, v_y=%f\n", 
+	LOG(("t=%d, p_id=%d, mass=%f, p_x=%f, p_y=%f, v_x=%f, v_y=%f\n",
 			time, (int)p[ID_COL], p[WEIGHT_COL], p[POS_X_COL], p[POS_Y_COL],
 			p[VOL_X_COL], p[VOL_Y_COL]));
 */
@@ -78,14 +78,14 @@ void print_particle(double *p, double time) {
 	if (p[WEIGHT_COL] <= 0.0) {
 		return;
 	}
-	LOG(("%lf,%d,%lf,%lf,%lf,%lf,%lf\n", 
+	LOG(("%d,%d,%lf,%lf,%lf,%lf,%lf\n",
 			time, (int)p[ID_COL], p[WEIGHT_COL], p[POS_X_COL], p[POS_Y_COL],
 			p[VOL_X_COL], p[VOL_Y_COL]));
 */
 	return;
 }
 
-void print_all_particles(double **P, int numParticle, double time) {
+void print_all_particles(double **P, int numParticle, int time) {
 	int i;
 	//LOG(("{\n"));
 	//LOG(("t,p_id,mass,p_x,p_y,v_x,v_y\n"));
@@ -102,10 +102,10 @@ void particles_gen(double **P, int light_cnt, int medium_cnt, int heavy_cnt, int
 	particles_gen_by_type(&P[light_cnt + medium_cnt + heavy_cnt], DUMMY,  padding_cnt);
 }
 
-void update(unsigned char* image, double **P, double **P_force, int total_p_cnt, 
+void update(unsigned char* image, double **P, double **P_force, int total_p_cnt,
 	int img_width, int img_height, double step_size, int regenerate_img) {
 	int i, cnt = 0;
-	
+
 	initilize_img(image, img_width, img_height);
 	for (i = 0; i < total_p_cnt; ++i) {
 		if (P[i][WEIGHT_COL] == DUMMY_WEIGHT) {
